refactor: remove ColorWrap HOC
