+++ src/layout/colorPicker/common/index.js
@@ -1,9 +1,8 @@
 import Alpha from './Alpha';
 import CheckBoard from './CheckBoard';
-import ColorWrap from './ColorWrap';
 import EditableInput from './EditableInput';
 import Hue from './Hue';
 import Saturation from './Saturation';
 import Swatch from './Swatch';
 
-export { Alpha, CheckBoard, EditableInput, Hue, Saturation, ColorWrap, Swatch };
+export { Alpha, CheckBoard, EditableInput, Hue, Saturation, Swatch };
